+++ server/src/main/java/com/dawn/server/service/CategoryService.java
@@ -7,11 +7,11 @@ import com.dawn.server.dto.CategoryDto;
 public interface CategoryService {
     List<CategoryDto> findAll();
 
-    CategoryDto findOne(final Long categoryId);
+    CategoryDto findOne(final Integer categoryId);
 
     CategoryDto save(final CategoryDto category);
 
-    CategoryDto update(final Long categoryId, final CategoryDto category);
+    CategoryDto update(final Integer categoryId, final CategoryDto category);
 
-    boolean deleteById(final Long categoryId);
+    boolean deleteById(final Integer categoryId);
 }
